# Conflicts:
#	src/main/webapp/WEB-INF/views/store/store_list.jsp
#	src/main/webapp/resources/css/storelist.css
#	src/main/webapp/resources/css/style.css

+++ src/main/java/com/spring/gogidang/domain/StoreVO.java
@@ -19,7 +19,7 @@ public class StoreVO {
 	private int delibery;
 	private int confirm;
 	
-	private int avgStar;
+	private Double avgStar;
 	
 	/* menu column*/
 	private int menu_num;
